+++ plugins/fengyun3_support/fengyun3/module_fengyun_mpt_decoder.h
@@ -29,8 +29,8 @@ namespace fengyun3
 
         std::ifstream data_in;
         std::ofstream data_out;
-        std::atomic<size_t> filesize;
-        std::atomic<size_t> progress;
+        std::atomic<uint64_t> filesize;
+        std::atomic<uint64_t> progress;
 
         viterbi::Viterbi1_2 viterbi1, viterbi2;
         deframing::BPSK_CCSDS_Deframer deframer;
